add development and production env files using mock api when in development mode

+++ src/lib/api.ts
@@ -1,9 +1,16 @@
 import * as mytype from "./type";
+import { faker } from "@faker-js/faker";
+import d from "./mock_data.json";
 
-const baseUrl = import.meta.env.VITE_API_URL;
+const mode = import.meta.env.MODE;
+const baseUrl = mode === "production" ? "" : import.meta.env.VITE_API_URL;
+const MockData = d as mytype.MockDataJson;
 
 export async function addDC(body: Pick<mytype.Datacenter, "name" | "height" | "ip_ranges">) {
-  console.log("addDC", body);
+  if (mode === "development") {
+    console.log("addDC", body);
+    return Promise.resolve(faker.string.uuid());
+  }
 
   const response = await fetch(`${baseUrl}/dc/`, {
     method: "POST",
@@ -19,7 +26,10 @@ export async function addDC(body: Pick<mytype.Datacenter, "name" | "height" | "i
 }
 
 export async function getAllDC(): Promise<mytype.SimpleDatacenter[]> {
-  console.log("getAllDC");
+  if (mode === "development") {
+    console.log("getAllDC");
+    return Promise.resolve(MockData.dc as mytype.SimpleDatacenter[]);
+  }
 
   const response = await fetch(`${baseUrl}/dc/all`);
   if (!response.ok) {
@@ -29,7 +39,14 @@ export async function getAllDC(): Promise<mytype.SimpleDatacenter[]> {
 }
 
 export async function getDC(dc_id: string): Promise<mytype.Datacenter> {
-  console.log("getDC", dc_id);
+  if (mode === "development") {
+    console.log("getDC", dc_id);
+    const dc = MockData.dc.find((dc) => dc.id === dc_id);
+    if (!dc) {
+      return Promise.reject(new Error("Datacenter not found"));
+    }
+    return Promise.resolve(dc);
+  }
 
   const response = await fetch(`${baseUrl}/dc/${dc_id}`);
   if (!response.ok) {
@@ -42,7 +59,15 @@ export async function modifyDC(
   dc_id: string,
   body: Partial<Pick<mytype.Datacenter, "name" | "height" | "ip_ranges">>,
 ): Promise<void> {
-  console.log("modifyDC", dc_id, body);
+  if (mode === "development") {
+    console.log("modifyDC", dc_id, body);
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+        reject(new Error(`Failed to edit datacenter with id ${dc_id}`));
+      }, 300);
+      resolve();
+    });
+  }
 
   const response = await fetch(`${baseUrl}/dc/${dc_id}`, {
     method: "PUT",
@@ -58,7 +83,15 @@ export async function modifyDC(
 }
 
 export async function deleteDC(dc_id: string): Promise<void> {
-  console.log("deleteDC", dc_id);
+  if (mode === "development") {
+    console.log("deleteDC", dc_id);
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+        reject(new Error(`Failed to delete datacenter with id ${dc_id}`));
+      }, 300);
+      resolve();
+    });
+  }
 
   const response = await fetch(`${baseUrl}/dc/${dc_id}`, {
     method: "DELETE",
@@ -70,7 +103,10 @@ export async function deleteDC(dc_id: string): Promise<void> {
 }
 
 export async function addRoom(body: Pick<mytype.Room, "name" | "height" | "dc_id">) {
-  console.log("addRoom", body);
+  if (mode === "development") {
+    console.log("addRoom", body);
+    return Promise.resolve(faker.string.uuid());
+  }
 
   const response = await fetch(`${baseUrl}/room/`, {
     method: "POST",
@@ -86,7 +122,14 @@ export async function addRoom(body: Pick<mytype.Room, "name" | "height" | "dc_id
 }
 
 export async function getRoom(room_id: string): Promise<mytype.Room> {
-  console.log("getRoom", room_id);
+  if (mode === "development") {
+    console.log("getRoom", room_id);
+    const room = MockData.room.find((room) => room.id === room_id);
+    if (!room) {
+      return Promise.reject(new Error("Room not found"));
+    }
+    return Promise.resolve(room);
+  }
 
   const response = await fetch(`${baseUrl}/room/${room_id}`);
   if (!response.ok) {
@@ -99,7 +142,15 @@ export async function modifyRoom(
   room_id: string,
   body: Partial<Pick<mytype.Room, "name" | "height" | "dc_id">>,
 ): Promise<void> {
-  console.log("modifyRoom", room_id, body);
+  if (mode === "development") {
+    console.log("modifyRoom", room_id, body);
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+        reject(new Error(`Failed to edit room with id ${room_id}`));
+      }, 300);
+      resolve();
+    });
+  }
 
   const response = await fetch(`${baseUrl}/room/${room_id}`, {
     method: "PUT",
@@ -115,7 +166,15 @@ export async function modifyRoom(
 }
 
 export async function deleteRoom(room_id: string): Promise<void> {
-  console.log("deleteRoom", room_id);
+  if (mode === "development") {
+    console.log("deleteRoom", room_id);
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+        reject(new Error(`Failed to delete room with id ${room_id}`));
+      }, 300);
+      resolve();
+    });
+  }
 
   const response = await fetch(`${baseUrl}/room/${room_id}`, {
     method: "DELETE",
@@ -129,7 +188,10 @@ export async function deleteRoom(room_id: string): Promise<void> {
 export async function addRack(
   body: Pick<mytype.Rack, "name" | "height" | "room_id" | "dc_id" | "service_id">,
 ) {
-  console.log("addRack", body);
+  if (mode === "development") {
+    console.log("addRack", body);
+    return Promise.resolve(faker.string.uuid());
+  }
 
   const response = await fetch(`${baseUrl}/rack/`, {
     method: "POST",
@@ -144,18 +206,15 @@ export async function addRack(
   return response.json();
 }
 
-export async function getAllRack(): Promise<mytype.SimpleRack[]> {
-  console.log("getAllRack");
-
-  const response = await fetch(`${baseUrl}/rack/all`);
-  if (!response.ok) {
-    return Promise.reject(new Error("Failed to fetch racks"));
-  }
-  return response.json();
-}
-
 export async function getRack(rack_id: string): Promise<mytype.Rack> {
-  console.log("getRack", rack_id);
+  if (mode === "development") {
+    console.log("getRack", rack_id);
+    const rack = MockData.rack.find((rack) => rack.id === rack_id);
+    if (!rack) {
+      return Promise.reject(new Error("Rack not found"));
+    }
+    return Promise.resolve(rack);
+  }
 
   const response = await fetch(`${baseUrl}/rack/${rack_id}`);
   if (!response.ok) {
@@ -168,7 +227,15 @@ export async function modifyRack(
   rack_id: string,
   body: Partial<Pick<mytype.Rack, "name" | "height" | "room_id" | "service_id">>,
 ): Promise<void> {
-  console.log("modifyRack", rack_id, body);
+  if (mode === "development") {
+    console.log("modifyRack", rack_id, body);
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+        reject(new Error(`Failed to edit room with id ${rack_id}`));
+      }, 300);
+      resolve();
+    });
+  }
 
   const response = await fetch(`${baseUrl}/rack/${rack_id}`, {
     method: "PUT",
@@ -184,7 +251,15 @@ export async function modifyRack(
 }
 
 export async function deleteRack(rack_id: string): Promise<void> {
-  console.log("deleteRack", rack_id);
+  if (mode === "development") {
+    console.log("deleteRack", rack_id);
+    return new Promise((resolve, reject) => {
+      setTimeout(() => {
+        reject(new Error(`Failed to delete room with id ${rack_id}`));
+      }, 300);
+      resolve();
+    });
+  }
 
   const response = await fetch(`${baseUrl}/rack/${rack_id}`, {
     method: "DELETE",
@@ -198,7 +273,10 @@ export async function deleteRack(rack_id: string): Promise<void> {
 export async function addHost(
   body: Pick<mytype.Host, "name" | "height" | "rack_id" | "room_id" | "dc_id" | "pos">,
 ): Promise<string> {
-  console.log("addHost", body);
+  if (mode === "development") {
+    console.log("addHost", body);
+    return Promise.resolve(faker.string.uuid());
+  }
 
   const response = await fetch(`${baseUrl}/host/`, {
     method: "POST",
@@ -215,7 +293,14 @@ export async function addHost(
 }
 
 export async function getHost(host_id: string): Promise<mytype.Host> {
-  console.log("getHost", host_id);
+  if (mode === "development") {
+    console.log("getHost", host_id);
+    const host = MockData.host.find((host) => host.id === host_id);
+    if (!host) {
+      return Promise.reject(new Error("Host not found"));
+    }
+    return Promise.resolve(host);
+  }
 
   const response = await fetch(`${baseUrl}/host/${host_id}`);
   if (!response.ok) {
@@ -228,7 +313,10 @@ export async function modifyHost(
   host_id: string,
   body: Partial<Pick<mytype.Host, "name" | "height" | "ip" | "status" | "rack_id" | "pos">>,
 ): Promise<void> {
-  console.log("modifyHost", host_id, body);
+  if (mode === "development") {
+    console.log("modifyHost", host_id, body);
+    return Promise.resolve();
+  }
 
   const response = await fetch(`${baseUrl}/host/${host_id}`, {
     method: "PUT",
@@ -244,7 +332,10 @@ export async function modifyHost(
 }
 
 export async function deleteHost(host_id: string): Promise<void> {
-  console.log("deleteHost", host_id);
+  if (mode === "development") {
+    console.log("deleteHost", host_id);
+    return Promise.resolve();
+  }
 
   const response = await fetch(`${baseUrl}/host/${host_id}`, {
     method: "DELETE",
@@ -257,8 +348,11 @@ export async function deleteHost(host_id: string): Promise<void> {
 
 export async function addService(
   body: Pick<mytype.Service, "name" | "n_racks" | "total_ip">,
-): Promise<mytype.Service> {
-  console.log("addService", body);
+): Promise<string> {
+  if (mode === "development") {
+    console.log("addService", body);
+    return Promise.resolve(faker.string.uuid());
+  }
 
   const response = await fetch(`${baseUrl}/service/`, {
     method: "POST",
@@ -270,11 +364,15 @@ export async function addService(
   if (!response.ok) {
     return Promise.reject(new Error("Failed to add service"));
   }
-  return response.json();
+  const data = await response.json();
+  return data.id as string;
 }
 
 export async function getAllService(): Promise<mytype.SimpleService[]> {
-  console.log("getAllService");
+  if (mode === "development") {
+    console.log("getAllService");
+    return Promise.resolve(MockData.service as mytype.SimpleService[]);
+  }
 
   const response = await fetch(`${baseUrl}/service/all`);
   if (!response.ok) {
@@ -284,7 +382,15 @@ export async function getAllService(): Promise<mytype.SimpleService[]> {
 }
 
 export async function getService(service_id: string): Promise<mytype.Service> {
-  console.log("getService", service_id);
+  if (mode === "development") {
+    console.log("getService", service_id);
+
+    const service = MockData.service.find((service) => service.id === service_id);
+    if (!service) {
+      return Promise.reject(new Error("Service not found"));
+    }
+    return Promise.resolve(service);
+  }
 
   const response = await fetch(`${baseUrl}/service/${service_id}`);
   if (!response.ok) {
@@ -297,7 +403,10 @@ export async function modifyService(
   service_id: string,
   body: Pick<mytype.Service, "name">,
 ): Promise<void> {
-  console.log("modifyService", service_id, body);
+  if (mode === "development") {
+    console.log("modifyService", service_id, body);
+    return Promise.resolve();
+  }
 
   const response = await fetch(`${baseUrl}/service/${service_id}`, {
     method: "PUT",
@@ -313,7 +422,10 @@ export async function modifyService(
 }
 
 export async function extendServiceRack(service_id: string, num: number): Promise<void> {
-  console.log("extendServiceRack", service_id, num);
+  if (mode === "development") {
+    console.log("extendServiceRack", service_id, num);
+    return Promise.resolve();
+  }
 
   const response = await fetch(`${baseUrl}/service/${service_id}/rack/extend`, {
     method: "PUT",
@@ -329,7 +441,10 @@ export async function extendServiceRack(service_id: string, num: number): Promis
 }
 
 export async function extendServiceIP(service_id: string, num: number): Promise<void> {
-  console.log("extendServiceIP", service_id, num);
+  if (mode === "development") {
+    console.log("extendServiceIP", service_id, num);
+    return Promise.resolve();
+  }
 
   const response = await fetch(`${baseUrl}/service/${service_id}/ip/extend`, {
     method: "PUT",
@@ -345,7 +460,10 @@ export async function extendServiceIP(service_id: string, num: number): Promise<
 }
 
 export async function deleteService(service_id: string): Promise<void> {
-  console.log("deleteService", service_id);
+  if (mode === "development") {
+    console.log("deleteService", service_id);
+    return Promise.resolve();
+  }
 
   const response = await fetch(`${baseUrl}/service/${service_id}`, {
     method: "DELETE",
